Move getter/setter for thread shared/local data to internal libs
