+++ src/ts_mruby.cpp
@@ -1,4 +1,3 @@
-#include <fstream>
 #include <iostream>
 #include <map>
 #include <pthread.h>
@@ -8,49 +7,10 @@
 #include <atscppapi/PluginInit.h>
 #include <atscppapi/RemapPlugin.h>
 
-#include "ts_mruby.hpp"
-#include "ts_mruby_init.hpp"
 #include "ts_mruby_internal.hpp"
-#include "ts_mruby_internal.hpp"
-#include "ts_mruby_request.hpp"
-#include "utils.hpp"
 
 using namespace std;
 using namespace atscppapi;
-
-namespace {
-
-// Global mruby scripts cache
-static MrubyScriptsCache *scriptsCache = NULL;
-
-// key specifying thread local data
-pthread_key_t threadKey = 0;
-
-// Initialize thread key when this plugin is loaded
-__attribute__((constructor)) void create_thread_keys() {
-  if (threadKey == 0) {
-    if (pthread_key_create(&threadKey, NULL) != 0) {
-      // XXX fatal error
-    }
-  }
-}
-
-// Note: Use pthread API's directly to have thread local parameters
-ThreadLocalMRubyStates *getMrubyStates() {
-  auto *state =
-      static_cast<ThreadLocalMRubyStates *>(pthread_getspecific(threadKey));
-
-  if (!state) {
-    state = new ThreadLocalMRubyStates();
-    if (pthread_setspecific(threadKey, state)) {
-      // XXX fatal error
-    }
-  }
-
-  return state;
-}
-
-} // anonymous namespace
 
 class MRubyPluginBase {
 protected:
@@ -58,7 +18,7 @@ protected:
 
   TSMrubyResult executeMrubyScript(Transaction &transaction) {
     // get or initialize thread local mruby VM
-    ThreadLocalMRubyStates *states = getMrubyStates();
+    auto* states = ts_mruby::getThreadLocalMrubyStates();
     mrb_state *mrb = states->getMrb();
 
     // get or compile mruby script
@@ -79,34 +39,6 @@ private:
   string filepath_;
   shared_ptr<TSMrubyContext> context_;
 };
-
-ThreadLocalMRubyStates::ThreadLocalMRubyStates() {
-  state_ = mrb_open();
-  ts_mrb_class_init(state_);
-}
-
-ThreadLocalMRubyStates::~ThreadLocalMRubyStates() {
-  mrb_close(state_);
-  state_ = NULL;
-}
-
-RProc *ThreadLocalMRubyStates::getRProc(const std::string &key) {
-  RProc *proc = procCache_[key];
-  if (!proc) {
-    const std::string &code = scriptsCache->load(key);
-
-    // compile
-    mrbc_context *context = mrbc_context_new(state_);
-    auto *st = mrb_parse_string(state_, code.c_str(), context);
-    proc = mrb_generate_code(state_, st);
-    mrb_pool_close(st->pool);
-
-    // store to cache
-    procCache_.insert(make_pair(key, proc));
-  }
-
-  return proc;
-}
 
 class MRubyPlugin : public GlobalPlugin, MRubyPluginBase {
 public:
@@ -143,11 +75,7 @@ void TSPluginInit(int argc, const char *argv[]) {
     RegisterGlobalPlugin(TS_MRUBY_PLUGIN_NAME, TS_MRUBY_PLUGIN_AUTHOR,
                          TS_MRUBY_PLUGIN_EMAIL);
 
-    if (!scriptsCache) {
-      scriptsCache = ts_mruby::utils::mockable_ptr<MrubyScriptsCache>();
-    }
-    scriptsCache->store(argv[1]);
-
+    ts_mruby::getInitializedGlobalScriptCache(argv[1]);
     new MRubyPlugin(argv[1]);
   }
 }
@@ -156,11 +84,7 @@ void TSPluginInit(int argc, const char *argv[]) {
 TSReturnCode TSRemapNewInstance(int argc, char *argv[], void **ih,
                                 char * /* ATS_UNUSED */, int /* ATS_UNUSED */) {
   if (argc == 3) {
-    if (!scriptsCache) {
-      scriptsCache = ts_mruby::utils::mockable_ptr<MrubyScriptsCache>();
-    }
-    scriptsCache->store(argv[2]);
-
+    ts_mruby::getInitializedGlobalScriptCache(argv[2]);
     new MRubyRemapPlugin(ih, argv[2]);
 
     return TS_SUCCESS;
