+++ src/ts_mruby_internal.cpp
@@ -3,12 +3,16 @@
 //
 */
 
+#include "ts_mruby_init.hpp"
 #include "ts_mruby_internal.hpp"
+#include "utils.hpp"
 
 #include <algorithm>
+#include <fstream>
+#include <iostream>
 #include <sstream>
-#include <string>
 
+#include <mruby/compile.h>
 #include <mruby/string.h>
 #include <ts/ts.h>
 
@@ -16,6 +20,18 @@ using namespace std;
 using namespace atscppapi;
 
 namespace {
+
+// key specifying thread local data
+pthread_key_t threadKey = 0;
+
+// Initialize thread key when this plugin is loaded
+__attribute__((constructor)) void create_thread_keys() {
+  if (threadKey == 0) {
+    if (pthread_key_create(&threadKey, NULL) != 0) {
+      // XXX fatal error
+    }
+  }
+}
 
 vector<string> split(const string &str, char delimiter) {
   vector<string> splitted;
@@ -32,6 +48,93 @@ vector<string> split(const string &str, char delimiter) {
 }
 
 } // anonymous namespace
+
+namespace ts_mruby {
+
+/*
+ * mruby scripts cache
+ */
+class MrubyScriptsCache {
+public:
+  MOCKABLE_ATTR
+  void store(const std::string &filepath) {
+    std::ifstream ifs(filepath);
+    const std::string code((std::istreambuf_iterator<char>(ifs)),
+                           std::istreambuf_iterator<char>());
+
+    scripts_.insert(make_pair(filepath, code));
+  }
+
+  MOCKABLE_ATTR
+  const std::string &load(const std::string &filepath) {
+    return scripts_[filepath];
+  }
+
+#ifdef MOCKING
+  using mock_type = class MrubyScriptsCacheMock;
+#endif // MOCKING
+
+private:
+  std::map<std::string, std::string> scripts_;
+};
+
+/*
+ * Global mruby scripts cache
+ */
+static MrubyScriptsCache *scriptsCache = NULL;
+MrubyScriptsCache* getInitializedGlobalScriptCache(const string& filepath) {
+  if (!scriptsCache) {
+    scriptsCache = ts_mruby::utils::mockable_ptr<MrubyScriptsCache>();
+  }
+  scriptsCache->store(filepath);
+  
+  return scriptsCache;
+}
+
+ThreadLocalMRubyStates::ThreadLocalMRubyStates() {
+  state_ = mrb_open();
+  ts_mrb_class_init(state_);
+}
+
+ThreadLocalMRubyStates::~ThreadLocalMRubyStates() {
+  mrb_close(state_);
+  state_ = NULL;
+}
+
+RProc *ThreadLocalMRubyStates::getRProc(const std::string &key) {
+  RProc *proc = procCache_[key];
+  if (!proc) {
+    const std::string &code = scriptsCache->load(key);
+
+    // compile
+    mrbc_context *context = mrbc_context_new(state_);
+    auto *st = mrb_parse_string(state_, code.c_str(), context);
+    proc = mrb_generate_code(state_, st);
+    mrb_pool_close(st->pool);
+
+    // store to cache
+    procCache_.insert(make_pair(key, proc));
+  }
+
+  return proc;
+}
+
+// Note: Use pthread API's directly to have thread local parameters
+ThreadLocalMRubyStates *getThreadLocalMrubyStates() {
+  auto *state =
+      static_cast<ThreadLocalMRubyStates *>(pthread_getspecific(threadKey));
+
+  if (!state) {
+    state = new ThreadLocalMRubyStates();
+    if (pthread_setspecific(threadKey, state)) {
+      // XXX fatal error
+    }
+  }
+
+  return state;
+}
+
+} // ts_mruby namespace
 
 bool judge_tls(const string &scheme) {
   if (scheme == "https") {
